Update ei.py to up-sample EI
In the compute stage of the EI added a check for the resample flag passed to the domain coordinate. If it is true, use a cubic spline to up-sample the B matrix to a frequency range corresponding to a uniform sampling between fmin and fmax at a resolution of 1/T where T is the duration of the lowest chirp mass signal.